pyproject.toml
[project]
name = "quiz-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "fastapi>=0.121.0",
    "html5lib>=1.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openpyxl>=3.1.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "playwright>=1.55.0",
    "plotly>=6.4.0",
    "pydantic[email]>=2.12.4",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "speechrecognition>=3.14.4",
    "uvicorn>=0.38.0",
]
